:hammer: refactor: outbox package 구조 변경
