--- src/main/java/com/example/harumeonglog/global/common/entity/enums/EventType.java
@@ -1,5 +1,0 @@
-package com.example.harumeonglog.global.common.entity.enums;
-
-public enum EventType {
-    FCM
-}
