+++ src/main/java/com/example/harumeonglog/global/outbox/entity/enums/EventType.java
@@ -1,0 +1,5 @@
+package com.example.harumeonglog.global.outbox.entity.enums;
+
+public enum EventType {
+    FCM
+}
